+++ include/zarr.types.h
@@ -163,15 +163,17 @@ extern "C"
         ZarrDataType data_type;
         bool multiscale;
         ZarrDownsamplingMethod downsampling_method;
+        uint32_t max_levels; /**< Maximum number of downsampled levels in the
+                                  pyramid. 0 means no limit. */
         const size_t* storage_dimension_order;
     } ZarrArraySettings;
 
     /**
      * @brief Settings for a field of view in a high-content screening (HCS)
      * well.
-     * @note @p array_settings->output_key must be @p NULL, because the path to
+     * @note @p array_settings->output_key must be NULL, because the path to
      * the array is fully specified by @p path. Validation will fail if it is
-     * non-@p NULL
+     * non-NULL.
      */
     typedef struct
     {